+++ persistence/sql/sql_process_store_common.go
@@ -159,6 +159,31 @@ func (p sqlProcessStoreImpl) publishToLocalQueue(
 	return true, nil
 }
 
+func (p sqlProcessStoreImpl) getDedupIdToLocalQueueMessageMap(ctx context.Context, processExecutionId uuid.UUID,
+	consumedMessages []persistence.InternalLocalQueueMessage,
+) (map[string]extensions.LocalQueueMessageRow, error) {
+	if len(consumedMessages) == 0 {
+		return map[string]extensions.LocalQueueMessageRow{}, nil
+	}
+
+	var allConsumedDedupIdStrings []string
+	for _, consumedMessage := range consumedMessages {
+		allConsumedDedupIdStrings = append(allConsumedDedupIdStrings, consumedMessage.DedupId)
+	}
+
+	allConsumedLocalQueueMessages, err := p.session.SelectLocalQueueMessages(ctx, processExecutionId, allConsumedDedupIdStrings)
+	if err != nil {
+		return nil, err
+	}
+
+	dedupIdToLocalQueueMessageMap := map[string]extensions.LocalQueueMessageRow{}
+	for _, message := range allConsumedLocalQueueMessages {
+		dedupIdToLocalQueueMessageMap[message.DedupId.String()] = message
+	}
+
+	return dedupIdToLocalQueueMessageMap, nil
+}
+
 func (p sqlProcessStoreImpl) updateCommandResultsWithConsumedLocalQueueMessages(
 	commandResults *xdbapi.CommandResults,
 	consumedMessages []persistence.InternalLocalQueueMessage,
@@ -184,4 +209,52 @@ func (p sqlProcessStoreImpl) updateCommandResultsWithConsumedLocalQueueMessages(
 	}
 
 	return nil
+}
+
+func (p sqlProcessStoreImpl) hasCompletedWaitUntilWaiting(commandRequest xdbapi.CommandRequest, commandResults xdbapi.CommandResults) bool {
+	// TODO: currently, only consider the local queue results
+
+	localQueueToMessageCountMap := map[string]int{}
+	for _, localQueueMessage := range commandResults.LocalQueueResults {
+		localQueueToMessageCountMap[localQueueMessage.QueueName] += 1
+	}
+
+	switch commandRequest.GetWaitingType() {
+	case xdbapi.ANY_OF_COMPLETION:
+		for _, localQueueCommand := range commandRequest.LocalQueueCommands {
+			if localQueueToMessageCountMap[localQueueCommand.QueueName] == int(localQueueCommand.GetCount()) {
+				return true
+			}
+		}
+		return false
+	case xdbapi.ALL_OF_COMPLETION:
+		for _, localQueueCommand := range commandRequest.LocalQueueCommands {
+			if localQueueToMessageCountMap[localQueueCommand.QueueName] < int(localQueueCommand.GetCount()) {
+				return false
+			}
+		}
+		return true
+	case xdbapi.EMPTY_COMMAND:
+		return true
+	default:
+		return true
+	}
+}
+
+func (p sqlProcessStoreImpl) completeWaitUntilWaiting(ctx context.Context, tx extensions.SQLTransaction,
+	stateRow *extensions.AsyncStateExecutionRowForUpdate, waitingQueues *persistence.StateExecutionWaitingQueuesJson, taskShardId int32) error {
+	waitingQueues.CleanupFor(persistence.StateExecutionId{
+		StateId:         stateRow.StateId,
+		StateIdSequence: stateRow.StateIdSequence,
+	})
+
+	stateRow.Status = persistence.StateExecutionStatusExecuteRunning
+
+	return tx.InsertImmediateTask(ctx, extensions.ImmediateTaskRowForInsert{
+		ShardId:            taskShardId,
+		TaskType:           persistence.ImmediateTaskTypeExecute,
+		ProcessExecutionId: stateRow.ProcessExecutionId,
+		StateId:            stateRow.StateId,
+		StateIdSequence:    stateRow.StateIdSequence,
+	})
 }
